rename/move client-only classes to client directory
