+++ src/CoAPNet.Server/CoapResource.cs
@@ -19,7 +19,7 @@ using System.Collections.Generic;
 using System.Net.Http.Headers;
 using System.Threading.Tasks;
 
-namespace CoAPNet
+namespace CoAPNet.Server
 {
     public abstract class CoapResource
     {
