+++ src/CoAPNet.Server/ICoapTransport.cs
@@ -1,12 +1,12 @@
 ﻿#region License
 // Copyright 2017 Roman Vaughan (NZSmartie)
-//  
+//
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
-// 
+//
 //     http://www.apache.org/licenses/LICENSE-2.0
-// 
+//
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -16,10 +16,15 @@
 
 using System.Threading.Tasks;
 
-namespace CoAPNet
+namespace CoAPNet.Server
 {
-    public interface ICoapHandler
+
+    public interface ICoapTransport
     {
-        Task ProcessRequestAsync(ICoapConnectionInformation connection, byte[] payload);
+        Task BindAsync();
+
+        Task UnbindAsync();
+
+        Task StopAsync();
     }
 }
