+++ tests/CoAPNet.Tests/CoreLinkFormatTests.cs
@@ -1,12 +1,12 @@
 ﻿#region License
 // Copyright 2017 Roman Vaughan (NZSmartie)
-//  
+//
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
-// 
+//
 //     http://www.apache.org/licenses/LICENSE-2.0
-// 
+//
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -21,6 +21,7 @@ using NUnit;
 using NUnit.Framework;
 
 using CoAPNet;
+using CoAPNet.Server;
 
 namespace CoAPNet.Tests
 {
@@ -69,7 +70,7 @@ namespace CoAPNet.Tests
                     Media = "none",
                     Title = "Outside Temperature",
                     TitleExt = "Primo Sensor",
-                    SuggestedContentTypes = 
+                    SuggestedContentTypes =
                     {
                         Options.ContentFormatType.TextPlain,
                         Options.ContentFormatType.ApplicationJson
